refactored :multiselect_client_interface
